% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_results.r
\name{aggregate_clusters}
\alias{aggregate_clusters}
\title{Aggregate the hierarchical clusters estimated}
\usage{
aggregate_clusters(dir, depth = NULL)
}
\arguments{
\item{dir}{String, a path to the directory in which are the subclusters to be aggregated}

\item{depth}{Positive Integer or NULL, maximum depth of the subclustering to plot. If NULL, the maximal depth of the clusters is assumed.}
}
\description{
Aggregate the hierarchical clusters estimated
}
\seealso{
Other results aggregation: \code{\link{aggregate_classification_probabilities}},
  \code{\link{aggregate_completed}}
}
\author{
Étienne Goffinet
}
\concept{results aggregation}
