% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtcomp_hierachical_launch.r
\name{launch_Mixtcomp_Hierarchical_predict}
\alias{launch_Mixtcomp_Hierarchical_predict}
\title{Launch the Mixtcomp prediction based on an existing hierarchical cluster}
\usage{
launch_Mixtcomp_Hierarchical_predict(data_path, param_dir,
  output_dir = NULL, mcStrategy = NULL)
}
\arguments{
\item{data_path}{data file path}

\item{param_dir}{A string defining the path of the directory in which to fint the parameters hierarchical tree on which the hierarchical predict will be based}

\item{output_dir}{(optional) if not null, the results will be written in this directory.
If Null, a directory with the name of the dataset will be created and used as output directory instead.}

\item{mcStrategy}{(optional) Additional parameters for mixtcomp launch list containing elements: (nbBurnInIter, nbIter, nbGibbsBurnInIter, nbGibbsIter)}
}
\description{
Launch the Mixtcomp prediction based on an existing hierarchical cluster
}
\examples{
\dontrun{
data_path <- "res/func/dat/dat.csv"
dataTestPath <- "res/func/dat/data_predict.csv"
descriptor_path <- "res/func/dat/desc.csv"

res <- launch_Mixtcomp_Hierarchical(data_path, descriptor_path, nClass = 3, depth = 5,
                                   output_dir = "res/func/out/",
                                   strategy = list(var = NULL, threshold_nInd = 10,
                                                   threshold_purity = 0.95), nCore = 3)


resPred <- launch_Mixtcomp_Hierarchical_predict(data_path = dataTestPath,
                                                param_dir = "res/func/out/dat/",
                                                output_dir = "res/func/outPred/"

# get partition
partition <- aggregate_clusters("res/func/outPred/data_predict/")

# get completed data
compl <- aggregate_completed("res/func/outPred/data_predict/", var = "Degradation_type")

}

}
\seealso{
\link{launch_Mixtcomp_Hierarchical} \link{aggregate_clusters}
}
\author{
Étienne Goffinet
}
