% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.r
\name{plot_categorical_hierarchique}
\alias{plot_categorical_hierarchique}
\title{Plot categorical data by clusters}
\usage{
plot_categorical_hierarchique(clusters, var, data, depth,
  dictionary = NULL, order_by = NULL, source = NULL)
}
\arguments{
\item{clusters}{vector of the clusters associated with data}

\item{var}{String, The name of the categorical data to plot}

\item{data}{the data to be plotted}

\item{depth}{Positive Integer, maximum depth of the subclustering to plot}

\item{dictionary}{(optional) List, the dictionary in which are stored the real names of the categorical data levels}

\item{order_by}{(optional) element among : size or purity, which define the order of the different cluster in the histogram.}

\item{source}{(optional) A string specifying the name of the source to be added to the plotly plot (in case the plot is used in a shiny app)}
}
\value{
a plotly plot
}
\description{
Plot categorical data by clusters
}
\seealso{
Other visualization: \code{\link{plot_functional_hierarchique}},
  \code{\link{plot_proportion_hierarchique}},
  \code{\link{plot_quality_pred_hierarchique}}
}
\author{
Étienne Goffinet
}
\concept{visualization}
