% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtCompHier-package.R
\docType{package}
\name{RMixtCompHier-package}
\alias{RMixtCompHier-package}
\title{RMixtCompHier}
\description{
Allow to launch MixtComp hierarchically, in both learn and predict mode.
}
\seealso{
\code{\link{launch_Mixtcomp_Hierarchical}}
}
\keyword{package}
